refactor(areaLeaderboard): split into components
