+++ src/components/leaderboard/AreaLeaderboardItemName.svelte
