+++ src/lib/comp.ts
@@ -10,10 +10,8 @@ export { default as AboutPlus } from '../components/AboutPlus.svelte';
 export { default as AboutTagger } from '../components/AboutTagger.svelte';
 export { default as AppCard } from '../components/AppCard.svelte';
 export { default as AreaActivity } from '../components/AreaActivity.svelte';
-export { default as AreaLeaderboard } from '../components/AreaLeaderboard.svelte';
-export { default as AreaLeaderboardItem } from '../components/AreaLeaderboardItem.svelte';
-export { default as AreaLeaderboardMinimal } from '../components/AreaLeaderboardMinimal.svelte';
-export { default as AreaLeaderboardSkeleton } from '../components/AreaLeaderboardSkeleton.svelte';
+export { default as AreaLeaderboard } from '../components/leaderboard/AreaLeaderboard.svelte';
+export { default as AreaLeaderboardItemName } from '../components/leaderboard/AreaLeaderboardItemName.svelte';
 export { default as AreaMap } from '../components/AreaMap.svelte';
 export { default as AreaMerchantHighlights } from '../components/AreaMerchantHighlights.svelte';
 export { default as AreaPage } from '../components/AreaPage.svelte';
